+++ test/with-momentjs/show_time_test.js
@@ -2,8 +2,8 @@ import React from 'react'
 // import ReactDOM from 'react-dom'
 // import TestUtils from 'react-addons-test-utils'
 import { mount } from 'enzyme'
-import DatePicker from '../src/datepicker.jsx'
-import TimeComponent from '../src/time'
+import DatePicker from '../../src/datepicker.jsx'
+import TimeComponent from '../../src/time'
 
 describe('DatePicker', () => {
   let sandbox
