+++ test/with-momentjs/week_number_test.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import WeekNumber from '../src/week_number'
+import WeekNumber from '../../src/week_number'
 import { shallow } from 'enzyme'
 
 function renderWeekNumber (weekNumber, props = {}) {
