+++ test/with-momentjs/year_dropdown_options_test.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import YearDropdownOptions from '../src/year_dropdown_options.jsx'
+import YearDropdownOptions from '../../src/year_dropdown_options.jsx'
 import { mount, shallow } from 'enzyme'
 
 describe('YearDropdownOptions', () => {
